Move base JsonRPC classes to separate package
The base classes for JSON RPC that (i.e. no bitcoinj dependencies) have
been moved to a package named com.msgilligan.jsonrpc — I want a cleaner
separation between the bitcoins-dependent parts of the code and the
code that has no dependency on bitcoinj. This code may be moved to a
separate submodule and JAR in the future.

This is a breaking change, but is easy enough to fix by just updating
the import statements.